+++ CHANGELOG.rst
@@ -1,9 +1,15 @@
 
 Changelog
 =========
+Unreleased changes
+------------------
+* Added a setting to enable and disable momentary cessation (MC_ENABLE)
+* Added settings to configure the momentary cessation voltage threshold (MC_LVRT_V1, MC_HVRT_V1)
+* Fixed a bug on momentary cessation not triggering on high voltage condition.
+
 2.1.3 (2023-12-05)
 ------------------
-* Add a property for SoC of BESS DER using .bess_soc
+* Added a property for SoC of BESS DER using .bess_soc
 * Automatically update the reactive capability curve if nameplate parameter is modified
   (when user defined curve is not provided)